Tidied up pagination reducer
