+++ src/app/store/reducers.module.ts
@@ -13,9 +13,9 @@ import { createAppReducer } from './reducers/create-application.reducer';
 import { dashboardReducer } from './reducers/dashboard-reducer';
 import { listReducer } from './reducers/list.reducer';
 import { requestPaginationReducer } from './reducers/pagination-reducer.generator';
-import { appMetaDataReducer } from './reducers/pagination-reducer/app-metadata-reducer.generator';
 import { uaaSetupReducer } from './reducers/uaa-setup.reducers';
 import { ActionReducerMap } from '@ngrx/store/src/models';
+import { appMetaDataReducer } from './reducers/app-metadata-reducer.generator';
 
 
 export function logger(reducer) {
