+++ src/app/store/reducers/app-metadata-reducer.generator.ts
@@ -1,7 +1,7 @@
+import { MetadataState } from '../types/app-metadata.types';
+import { appMetadataReducer } from './app-metadata.reducer';
+import { appMetadataRequestReducer } from './app-metadata-request.reducer';
 import { ActionReducerMap, combineReducers } from '@ngrx/store';
-import { MetadataState } from '../../types/app-metadata.types';
-import { appMetadataReducer } from '../app-metadata.reducer';
-import { appMetadataRequestReducer } from '../app-metadata-request.reducer';
 
 const appMetadataReducers: ActionReducerMap<MetadataState> = {
   values: appMetadataReducer,
