+++ src/app/store/reducers/pagination-reducer/pagination-reducer-set-page.ts
@@ -1,0 +1,11 @@
+import { SetPage } from '../../actions/pagination.actions';
+import { PaginationEntityState } from '../../types/pagination.types';
+
+export function paginationSetPage(state: PaginationEntityState, action: SetPage) {
+  return {
+    ...state,
+    error: false,
+    currentPage: action.pageNumber
+  };
+}
+
