+++ src/app/store/reducers/pagination-reducer/pagination-reducer-start.ts
@@ -1,0 +1,9 @@
+export function paginationStart(state) {
+  return {
+    ...state,
+    fetching: true,
+    error: false,
+    message: '',
+  };
+}
+
