+++ src/app/store/reducers/pagination-reducer/pagination-reducer.failure.ts
@@ -1,0 +1,9 @@
+import { PaginationEntityState } from '../../types/pagination.types';
+export function paginationFailure(state: PaginationEntityState, action) {
+  return {
+    ...state,
+    fetching: false,
+    error: true,
+    message: action.message
+  };
+}
